inject script and content script added
